Refactor out SignUpCard component to common components to be reused for SignIn
